pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "edge-connect-face-reconstruction"
authors = [
    {name = "ABDULLAH AHMAD", email = "your.email@example.com"},
]
description = "EdgeConnect-inspired face reconstruction for inpainting masked facial regions"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
keywords = ["computer-vision", "image-inpainting", "face-reconstruction", "edge-connect", "deep-learning"]
dependencies = [
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.0.0",
    "pillow>=8.0.0",
    "opencv-python>=4.5.0",
    "scikit-image>=0.18.0",
    "pyyaml",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "isort>=5.0",
    "mypy>=0.800",
    "bandit>=1.7.0",
    "safety>=1.10.0",
    "pre-commit>=2.0.0",
]
gpu = [
    "torch>=1.9.0+cu111",
    "torchvision>=0.10.0+cu111",
]
jupyter = [
    "jupyter>=1.0",
    "notebook>=6.0",
    "ipywidgets>=7.6.0",
]

[project.urls]
Homepage = "https://github.com/ABDULLAH-AHMAD-OFFICIAL/edge-connect-face-reconstruction"
Documentation = "https://github.com/ABDULLAH-AHMAD-OFFICIAL/edge-connect-face-reconstruction/wiki"
Repository = "https://github.com/ABDULLAH-AHMAD-OFFICIAL/edge-connect-face-reconstruction.git"
"Bug Reports" = "https://github.com/ABDULLAH-AHMAD-OFFICIAL/edge-connect-face-reconstruction/issues"
Changelog = "https://github.com/ABDULLAH-AHMAD-OFFICIAL/edge-connect-face-reconstruction/blob/main/CHANGELOG.md"

[project.scripts]
face-reconstruct = "scripts.process_user_image:main"
face-test = "scripts.test_reconstruction:main"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.dynamic]
version = {attr = "src.__version__"}

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
    ".venv",
    ".tox",
]

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests as requiring GPU",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]